Moved files to be less misleading
