+++ src/components/App/Menu/Skills.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { ReduxStore } from '../../../ts-interfaces/interfaces';
+import { ReduxStore } from 'Interfaces/interfaces';
 
 const Skills = () => {
   const skills = useSelector((state: ReduxStore) => state.aboutNathan.skills);
